+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/FLTClusterManagersController.m
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import "ClusterManagersController.h"
+#import "FLTClusterManagersController.h"
 #import "FLTGoogleMapJSONConversions.h"
 #import "GoogleMarkerUtilities.h"
