+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/google_maps_flutter_ios-umbrella.h
@@ -6,7 +6,7 @@
 #import <google_maps_flutter_ios/FLTGoogleMapJSONConversions.h>
 #import <google_maps_flutter_ios/FLTGoogleMapTileOverlayController.h>
 #import <google_maps_flutter_ios/FLTGoogleMapsPlugin.h>
-#import <google_maps_flutter_ios/GMSMarker+Userdata.h>
+#import <google_maps_flutter_ios/GoogleMarkerUtilities.h>
 
 FOUNDATION_EXPORT double google_maps_flutterVersionNumber;
 FOUNDATION_EXPORT const unsigned char google_maps_flutterVersionString[];
